requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
strawberry-graphql[fastapi]==0.200.0
pytest-cov==4.1.0
